post, comment 로그인 권한부여 완료

+++ routes/comments.js
@@ -3,6 +3,7 @@ const Mongoose = require('mongoose')
 const Comments = require('../models/comments')
 const Posts = require('../models/posts')
 const User = require('../models/user')
+const authMiddleware = require('../middlewares/auth-middleware')
 
 const router = express.Router()
 
@@ -19,21 +20,21 @@ router.get('/:post_id/comments', async (req, res) => {
 })
 
 //댓글 작성
-router.post('/:post_id/comments', async (req, res) => {
+router.post('/:post_id/comments', authMiddleware, async (req, res) => {
     const {post_id} = req.params
     console.log(post_id)
     const { content } = req.body
-    // const { user } = res.locals
-    // const userId = user.userId
+    const { user } = res.locals
+    const nickname = user.nickname
     let date = new Date().toISOString()
   
     try {
       const _id = new Mongoose.Types.ObjectId()
-      await Comments.create({ _id, content, date })
-    //   await User.findOneAndUpdate(
-    //     { userId: userId },
-    //     { $push: { comment_id: _id } }
-    //   )
+      await Comments.create({ _id, nickname, content, date })
+      // await User.findOneAndUpdate(
+      //   { nickname: nickname },
+      //   { $push: { comment_id: _id } }
+      // )
       await Posts.findOneAndUpdate(
         { _id: post_id },
         { $push: { comment_id: _id } }
@@ -49,38 +50,38 @@ router.post('/:post_id/comments', async (req, res) => {
 })
 
 //댓글 수정 
-router.patch('/:post_id/comments/:comment_id', async (req, res) => {
+router.patch('/:post_id/comments/:comment_id', authMiddleware, async (req, res) => {
     const { post_id, comment_id } = req.params
     console.log(post_id, comment_id)
     const { content } = req.body
-    // const { user } = res.locals
-    // const userId = user.userId
-    // const c_userId = await Comments.findOne({ _id: comment_id })
+    const { user } = res.locals
+    const nickname = user.nickname
+    const c_nickname = await Comments.findOne({ _id: comment_id })
     let date = new Date().toISOString()
+    console.log(c_nickname, nickname)
   
-    // if (c_userId.userId === userId) {
+    if (c_nickname.nickname === nickname) {
       await Comments.updateMany({ _id: comment_id }, { $set: { content, date } })
       res.send({ result: 'success' })
-    // } else {
-    //   res.send({ msg: '내 댓글이 아닙니다.' })
-    // }
+    } else {
+      res.send({ msg: '내 댓글이 아닙니다.' })
+    }
 })
 
 //댓글 삭제
-router.delete('/:post_id/comments/:comment_id', async (req, res) => {
+router.delete('/:post_id/comments/:comment_id', authMiddleware, async (req, res) => {
     const { post_id, comment_id } = req.params
-    // const { comment_id } = req.body
-    // const { user } = res.locals
-    // const userId = user.userId
+    const { user } = res.locals
+    const nickname = user.nickname
     const c_userId = await Comments.findOne({ _id: comment_id })
   
-    // if (c_userId.userId === userId) {
+    if (c_userId.nickname === nickname) {
       await Posts.findOneAndUpdate({ _id: post_id }, { $pull: { comment_id: comment_id } })
       await Comments.deleteOne({ comment_id })
       res.send({ result: 'success' })
-    // } else {
-    //   res.send({ msg: '내 댓글이 아닙니다.' })
-    // }
+    } else {
+      res.send({ msg: '내 댓글이 아닙니다.' })
+    }
 })
 
 module.exports = router;
